+++ extensions/web3-profile/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Rainbow Profile Changelog
 
+## [Bug Fixes] - 2022-09-06
+
+- Fix "memory leak" caused by ENS Avatar Records using Zora Contract [#2816](https://github.com/raycast/extensions/issues/2816)
+
 ## [Initial Version] - 2022-09-05